requirements.txt
streamlit
langchain
chromadb
sentence-transformers
transformers
faiss-cpu
PyPDF2
python-docx
langchain-community
python-dotenv
huggingface-hub


streamlit-webrtc
pyaudio 
# av   
pyttsx3
ffmpeg-python
sounddevice
pydub
speechrecognition
openai-whisper
websockets
requests

docx2txt
pdfplumber
unstructured[docx,pdf]

tqdm
numpy
pandas
scipy

httpx
Groq
langchain-groq
langchain-together

huggingface-hub
langchain-huggingface
